Rename things in the inbox
