+++ web/src/index.js
@@ -16,10 +16,10 @@ import PatientOverview from './ehr/patient/PatientOverview.vue'
 import NewPatient from './ehr/patient/NewPatient.vue'
 import EditPatient from "./ehr/patient/EditPatient.vue"
 import NewDossier from "./ehr/patient/dossier/New.vue"
-import NewTransfer from "./ehr/patient/transfer/NewTransfer.vue"
-import EditTransfer from "./ehr/patient/transfer/EditTransfer.vue"
-import TransferRequest from "./ehr/transfer/TransferRequest.vue"
-import Inbox from "./ehr/Inbox.vue"
+import NewTransfer from "./ehr/transfer/NewTransfer.vue"
+import EditTransfer from "./ehr/transfer/EditTransfer.vue"
+import TransferRequest from "./ehr/inbox/TransferRequest.vue"
+import Inbox from "./ehr/inbox/Inbox.vue"
 import Settings from "./ehr/Settings.vue"
 import Components from "./Components.vue"
 import Elevation from "./components/auth/SessionElevation.vue"
